requirements.txt

Flask
openai>=1.0.0
moviepy
openai-whisper
torch
